refactor: ParentHome Summary JSX 로직 함수로 분리
