+++ src/components/home/KidList.tsx
@@ -1,4 +1,3 @@
-import { IKid } from '@store/slices/familySlice';
 import { Dispatch, SetStateAction } from 'react';
 import styled from 'styled-components';
 import { ReactComponent as UsernameUnderline } from '@assets/borders/username-underline.svg';
