+++ src/lib/types/IKid.ts
@@ -1,0 +1,8 @@
+import { TLevel } from './TLevel';
+
+export interface IKid {
+  username: string;
+  isFemale: boolean;
+  level: TLevel;
+  kidId: number;
+}
